refactor: Add OfferPreviewList.
