+++ src/components/offer-preview-list/index.ts
@@ -1,0 +1,2 @@
+import OfferPreviewList from './offer-preview-list';
+export default OfferPreviewList;
