+++ src/components/offer-reviews/offer-review-item.tsx
@@ -1,4 +1,4 @@
-import { Review } from '../../types/offer';
+import { Review } from '../../types/review';
 import { getRatingWidth } from '../../utils';
 
 type ReviewItemProps = {
